+++ src/ws_server/updateWinners.ts
@@ -1,0 +1,5 @@
+import { Winner } from './db';
+
+export const updateWinnersResponse = (winners: Winner[]) => {
+  return { type: 'update_winners', data: JSON.stringify(winners), id: 0 };
+};
